add 增加ehcache缓存

+++ src/main/java/com/zhaodesong/Edocumentsystem/po/Project.java
@@ -18,7 +18,7 @@ public class Project implements Serializable {
 
     private String name;
 
-    private Long createAccount;
+    private Integer createAccount;
 
     private LocalDateTime createTime;
